feat(#84): [Pooja|Tilak] remove old nearby state machines and refactor selectors
