+++ machines/openIdBle/scan/machine.ts
@@ -9,29 +9,31 @@ import {
   EventFrom,
   send,
   spawn,
-  StateFrom,
 } from 'xstate';
 import { createModel } from 'xstate/lib/model';
 import { EmitterSubscription, Linking, Platform } from 'react-native';
-import { DeviceInfo } from '../../components/DeviceInfoList';
+import { DeviceInfo } from '../../../components/DeviceInfoList';
 import { getDeviceNameSync } from 'react-native-device-info';
-import { VC, VerifiablePresentation } from '../../types/vc';
-import { AppServices } from '../../shared/GlobalContext';
-import { ActivityLogEvents, ActivityLogType } from '../activityLog';
-import { MY_LOGIN_STORE_KEY, VC_ITEM_STORE_KEY } from '../../shared/constants';
-import { offlineSubscribe } from '../../shared/openIdBLE/walletEventHandler';
+import { VC, VerifiablePresentation } from '../../../types/vc';
+import { AppServices } from '../../../shared/GlobalContext';
+import { ActivityLogEvents, ActivityLogType } from '../../activityLog';
+import {
+  MY_LOGIN_STORE_KEY,
+  VC_ITEM_STORE_KEY,
+} from '../../../shared/constants';
+import { offlineSubscribe } from '../../../shared/openIdBLE/walletEventHandler';
 import {
   check,
   PERMISSIONS,
   PermissionStatus,
   RESULTS,
 } from 'react-native-permissions';
-import { checkLocation, requestLocation } from '../../shared/location';
+import { checkLocation, requestLocation } from '../../../shared/location';
 import { CameraCapturedPicture } from 'expo-camera';
 import { log } from 'xstate/lib/actions';
-import { isBLEEnabled } from '../../lib/smartshare';
-import { createQrLoginMachine, qrLoginMachine } from '../QrLoginMachine';
-import { StoreEvents } from '../store';
+import { isBLEEnabled } from '../../../lib/smartshare';
+import { createQrLoginMachine, qrLoginMachine } from '../../QrLoginMachine';
+import { StoreEvents } from '../../store';
 import { WalletDataEvent } from 'react-native-openid4vp-ble/lib/typescript/types/bleshare';
 
 const { wallet } = openIdBLE;
@@ -84,7 +86,7 @@ const model = createModel(
       LOCATION_DISABLED: () => ({}),
       LOCATION_REQUEST: () => ({}),
       UPDATE_VC_NAME: (vcName: string) => ({ vcName }),
-      STORE_RESPONSE: (response: unknown) => ({ response }),
+      STORE_RESPONSE: (response: any) => ({ response }),
       APP_ACTIVE: () => ({}),
       FACE_VALID: () => ({}),
       FACE_INVALID: () => ({}),
@@ -95,6 +97,7 @@ const model = createModel(
   }
 );
 const QR_LOGIN_REF_ID = 'QrLogin';
+export const ScanEvents = model.events;
 
 export const scanMachine =
   /** @xstate-layout N4IgpgJg5mDOIC5SwMYEMB2BiAygYQCUBRIgOQH0AhAGQFUCBtABgF1FQAHAe1gEsAXXlwzsQAD0QAWAEwAaEAE9E0gOzSAdNICsTbaq0BOLdoAcAXzPzUmXIRIUAYgHk8tHMzZIQ3PoOGiJBBl5JQQTAwN1HSYARgMmADYDSS0TLS0LK3RsGiJyIgICJ0ZWUR8BIREvQODFRBjpEyimXSZJBNTdE2lpTJBrDHUUAAswFABrXgwoSgAbAFcwfi4ufmGcMAAnADdeFDAh0YmpqCwaWiIAFScnS4AJcgBBamonAHUiABEPMp4K-2qiC0kkiagSCQAzCYIZIVAkVLCQogjJJ1C1dNIYroIkwVH0BocxpNpnNFstVustrt9oTjtMznQrjd7uRPmQAJJfH5ecp+KqgQIGCEQ9RxLQQmIqcVQ1JyOphaQJZq6GRSmKSKH47K04kzBZLFZrDY7PYHTZgACOi1ggnp5yZtweZEeuW+pR5fz5ASBIPUYMh0Nh8MR8pUTC0fp02niMlSCRiWswOpOpINFON1LNlutttO9uujtZ7JwLuoXPdnE9lW9CCFIrFEqlwrS3SRQQ6ysV3RUJjUYYyln62pGRJT+vJRqppvUYAwaAARrNIFhuZXfNXAbXhaKjI3pS25aFeyo0S1MbEkuqYglE4NF2TDcNPrPeMvHgAFd9PPCXdkANSIVdvCrAEBWUOJ1GhQwYTaeIJSMNtJBiCN0QhFoTHjXQYVvdR7zTNZnwwV8ICwABxW4nHIHArl-UhSPcCtgPXUDxEQCFpFiTQYUaMMNVhHo23YppDDULRkJ0GFYRwlAlzQTYTjwYQMDGPksE+Ys8CcUhSCIH8gN5DcwIVQSTHDTRoRiBomHYrR4QhaTZPk6ZFIwZSUFUsQbTQfgDjQAAzHzNgACjZHBLiKABNchfwAWSIJxaEuABKLACRksA5IUpSVMqfSQP5Vja0PNipSYdR4wSSQmGSHQb0HAk-KmCAstcnLhFsR5SDy5iCsCcUIxiISEhMNIsRSLREJVcqWkkNIqq0FQYm6HDGowZrnOy9zKg6rqYk8Nd-l6n0BqGkaxLadJBIRMrewlUb42hSEVqalq3NU-BOoYaR9qYw6a2BE7GmG0aLom+UoWhP0NWkGEVGSAwEl6ertVgYYuAAdwARU2aguCgKYsAgYQDimbYuHGA5sdx-GMG6v7N26EVMRGxIhQMJaVBUK7OcgkFoXY6yxLqrIk1RjGqbxgn1JwGLiwYn6DJYvqIUBjCzrGy7waF9QVZaSVGjPNocLFrGcclwYbS4JzTjC4o8mIHB3y06i6a9TdLIiSCVaFcNrIhOGEiu8UdZUO6haWiEnuRpMUE23MsE07TdMucsFfymtEcE68yt1phukkAv43s6PBlj1qtumdQpnfTY8fNWBYCwTz+G83yAq2ILE50n92S06L2TihLktS4c45OKuMBruu4FgV3DMKgxFXUWb2mBAw1A1UPJo7btmfSRG4WksfK8EABbMAuHmfgE86vAiGoOelcQTPwfPSDoj5xHGxw81djAdHx6wDAEuCuUA-woCwLQd8nxHgp3IMQR4OAtKPyOggdUPQdZ2W6KoFI7RA7ynEhoZCSRMRxGGhxYuItBi-1fAAyuQCQG5nAWpDSWlu6XBQTWJaC1RSxCFhxW6ec2yqFMuVdo3YEaxlUD-MAf86FQHUAwnK0xmHUTLD+cgf48CcPdhqJUWJbLIVhD2RaXMCHjT9NZeEI0Ei4h7OYEu6gaH-0AcA5RYCIEAQIOyBwUVOqfG-Hfd8lx4FEExhcMKOijLDRPMCXisFpBIVsm2a88YdwwnZhqVIDQHFUKcbI2hrjGEnGYY8PAQSQnEHCUQSJjFFaoJiVEWEbQElJPwaEZC6pNBVUaNeboiRclDiTM4+Rii3GgOYXgW+98omFUGmY0IiSUg6wRqY2auJcTCyGdQgpLj6HjKYRA64pFSJlnIG4Ao5BE7UVIBwup6d3YQkXpBboERtDpAwmDDp3tNA6D0cYWEkcDAyLkePbI+xZizBOCue5PUuEmFhLwzpiM86hyEQQ6qN1hQLXPADGGILCn7LWiUiB0su7J1mYEJaiKsTIoEWikwwi0WinEbY4EHQV4Er2QooBxKVGeLwFRMgdy05ws3P7GI6gIjSvVLEaqGEFnKDhJEcIw0qpYjSItSh2z8mgqJetDxWAtGBKIME1OvwxVGUspHTQCMhRqG0M2Rl8pVD+xWcNdIlkKpbIJCM1xfLDXGuIAAKWTuaj0lq5kNCVNIO1-sej7mdYsiqYjwTDVsWCaRji-X6pJRPKeUB66N2bq3dQ-lApBRwHcR43i6L90HolFKvrdmjN5Qa8B+ba6FpnpSoEec-QmAaDoYdQovnKFGiysMyEOKtGsly1ts520oHUGfC+V8b6kDvg-WF9MrVettUkeNjqIbCNsvo9oU6lpGGgvO8eaAUD7A4D5Ei0tZY4Hlha3dcyQQniEteKdfylpMrdU8j1YlrypNvZXc0AArFSy5X1y17WghFJ5aX8NRWhJNygU2VRibK0O6ooMKO2FsXgfkFAnHZBAWcgh+AKCwA4MpeQ-zPHZG6UVX6qWSmERhes4iYaDoRPi7NLbx6kfkhRqjNGMB0YY0xu+5B2SkFY9QdjyHGlxJaSCRJxCUnxA0IelIZ0VaSmI+oCT5HKPTGo7RgQDGpmbpmTut20TexNPiTptpKTMLlVHYk2Ig7qqSHMygc03kSkcEJsTCeZMKaKO1DmhRYWMpMI4AgUmXB0B8g8MhhoiqEDaB7H54UEQEWQjEqF8LaWsBbFrpsdQHBZjeT8lbU+CXhlicriliLKj0uZey7lVgyHGgiiSP1BaecYY2RSekSIHQ7JGHjHDMzom9UKNJmgKFEBbOyfsywmWSGXPz0CKYcqiT14AtiBxAr15g6WQvCrIajRzObe27tuTWBiDhSil4nx7Ipm-mQcdp+hXhrnZ-Vdqyt2eySuSLDMShcOU4WaqgOOCHixvo-RGrj4FQ47mMAjcU7N-bAd-QjNI4HvU4WGJgCAUKSRLiIJserB2sfIcjqicEVVJHhl7LGsdaDGjKj1pq9e4IcJvd4C+zHR3OOucKmoRCDRJXhnjZzJCh9HEjjpFAXGg3hAZmnDr3UOAW78HmI3V4gPe4UGdK6PL7RIi7ksvzYzkg2w+y9hk2y6QuxIzySbk4+uIuG6nDSIP0wzfeUt1ga3sDbdFhLA7kHqD1RJCleKV3wp3dthMcvVI4o0hLdMgOQPRxdQh75Ebmk5orRwFzJcLgRA5yLjAHHlwCe+727LBxz9CuqUGBGsvf2OmOiJBhm2ILaIoIe3aEhSq0kK-B6y6HjANesz15tCcJvLeFxLg7zbvu0tSzhoOgP+oQ+mibzH7VSfoZDBRCggXYwvYclL9HNMKvlQN-JmmO+LYU+XgBubaePIHO3UgU-PvHHC-QrdeNDWaBETmIWReT3KEEOWqcICGR1D-XXb-MPE0CPZff-QA4AvgdqMAxPE-FPeXE7ZQMSEUCfDoR1aMdpRAQdJURGMSDCRGaVReXAyvVfavcPA4VHffZcSgvuKpCJEVfvOgtBCIOHWNMMFESyBaQXFQmfYUBECUaqBaELbXYgvXIQn-EQ9QGTYiLAD8L8MpX8ACEbBaJobnJCdieMKqDoPPcEUUPQOITEHoJbAQlfA3dfMwiwiQzvcA0JapWpWg0HVQWNHWbgvOVIBEL1PPJaUURaZIGQNmBaPEPoDALgGjeALwAYOQ0HAAWjYIQCqMlznC2lI3KNQVITEUp0qgwklEHWEV9G0EC0emBjPECMZwfHTBEKaJrA4iVCgiFCqhBGsjFDbGyVPFaCQg1CeXyPL0-z1BGMnEIIOEjygHGM3G0FRGmJgjmPgkF3ZiVFKn0FMhhEhADx1QONTAnEpD2PyS31zCOKMglElVh30HXhaThDbHBGEj1h7AqixDhiGO2PwneMzBnFbyXAgB+MVzSEgl6VxH9nXksmKlrERV6J6Dwy6HXhwjwjeMImIjRO43iCfyHwLmh3VEQjQjRCjHZjzkhBhDL2eMclejagKnqRrHYkSV4XXhSCxANghBSRVnKjBKhGlUWk12egDRcjenniFPFUSGXmZieTQkji9UFwlQGjKwdSHUGh9RRjRlNmpimBpMQGQM0CWlMkPXZmPEEg+V5jKyd05kSWNmtIlhpirnpzAHtLCB4Sv0jhZkMQRCugggyU+R7GySzTyRNkDKmEUWWGtjDKH0lSKzf25xGmvCDhFAySlGwWmIMNTIDLNiDLnF2CgF6ygCbzuC4HPjDMGk9lHyHyhEsjhgMG5imL5mJPPHDCrOeOPkOJgPkIRBuOsljULilClEEi91cNMnDGITSCPnLlzE7WngbjDLhlRDDHYmyPhHUO3lONUCxA6FDlDmBW10nJXV4HPkvn4A7L6UghyVWUXmxUvMxMWkSAWn9ieWIzDJBBSX1kjAvEXkSHcMtM63WzGWKX5XAqMCRQDkbEXnZh82DkwmvCeyBhezW0JWS0wAhQZynPP3kIaCQjZPEW6EqkslqEWRf0glWUlHWTDC1zySSzGQDXATDPYhFD4MyTlSH3hGESLlTUI3FySALnMzbTzWri7SLSEsWilT4NlSxAkoK0aB0BZSSHnMlHBH9kUsXTzVXTfJzI0qhEaHXlVThChF4wMvTzgkxDhEhA2J1T4qUv5X4vfOnNBx0CaABOHXDFHSZVmkMseL4kqgWnM3vUfWfQ7Kw2XkwneUJxJyZS6VAxlAaBhElB5ObSQtg3g1RKCrTwXKiCFAskjmGhVkF0VElBivTVxEVBTJ8q6wUXrN4EbMby4FbPbMqq4URg0HzMXh-XhHxNdRPFVQLiMGzgq3M0sykxsxkzkxzPXj9CBN9kYsWmqMlB5mqmEr+VG21RKtIqGGq0ixstmw1BeWbFSLiDJJIu5Viy22lw+3szuoIXZhPDhnOhf0YrzhR2ALLjekgDDI4g4lPE5m6CKpSOqNmvYrAy9Ug0cVpzWkormDAGZ3qw7PCGaEhLznVXsUOoSPRFFwWjkrqO2E+oquotBzhGPPnLPKXKNJGhVUVFVB-V7EGTSiMPwJCL2I-LSQoWqgXJnXv1CERklVSFKx7Pzgl0MK2OFt-wOOjwtxKKZqquqk0GskluSGlulPlCwKf0araEaHznHMFrVpMIIMRLrxzB32b2RNDJGvdkvG6Q8tjSwNDlNtCBZoL01VskNtsQQtLiFodpFsRIOIAM2CAJAJYk1KMh6H1v9EQLHJlCn2eVKjzivQBhVs2LwJjt-zELb0Zt+lgPTtBEVHaP7A1DSDz2BE7FSSKzGthPVtCJfChs9rTp7BFF7Cz38KskF0HXm20CWnBFjWlSRgsCAA */
@@ -102,7 +105,7 @@ export const scanMachine =
     {
       predictableActionArguments: true,
       preserveActionOrder: true,
-      tsTypes: {} as import('./scan.typegen').Typegen0,
+      tsTypes: {} as import('./machine.typegen').Typegen0,
       schema: {
         context: model.initialContext,
         events: {} as EventFrom<typeof model>,
@@ -856,103 +859,4 @@ export function createScanMachine(serviceRefs: AppServices) {
     ...scanMachine.context,
     serviceRefs,
   });
-}
-
-type State = StateFrom<typeof scanMachine>;
-
-export function selectReceiverInfo(state: State) {
-  return state.context.receiverInfo;
-}
-
-export function selectReason(state: State) {
-  return state.context.reason;
-}
-
-export function selectVcName(state: State) {
-  return state.context.vcName;
-}
-
-export function selectSelectedVc(state: State) {
-  return state.context.selectedVc;
-}
-
-export function selectIsScanning(state: State) {
-  return state.matches('findingConnection');
-}
-
-export function selectIsConnecting(state: State) {
-  return state.matches('connecting.inProgress');
-}
-
-export function selectIsConnectingTimeout(state: State) {
-  return state.matches('connecting.timeout');
-}
-
-//TODO: post discussion with team remove the selectIsExchangingDeviceInfo & selectIsExchangingDeviceInfoTimeOut functions
-export function selectIsExchangingDeviceInfo() {
-  return true;
-}
-
-export function selectIsExchangingDeviceInfoTimeout() {
-  return true;
-}
-
-export function selectIsReviewing(state: State) {
-  return state.matches('reviewing');
-}
-
-export function selectIsSelectingVc(state: State) {
-  return state.matches('reviewing.selectingVc');
-}
-
-export function selectIsSendingVc(state: State) {
-  return state.matches('reviewing.sendingVc.inProgress');
-}
-
-export function selectIsSendingVcTimeout(state: State) {
-  return state.matches('reviewing.sendingVc.timeout');
-}
-
-export function selectIsAccepted(state: State) {
-  return state.matches('reviewing.accepted');
-}
-
-export function selectIsRejected(state: State) {
-  return state.matches('reviewing.rejected');
-}
-
-export function selectIsInvalid(state: State) {
-  return state.matches('invalid');
-}
-
-export function selectIsBluetoothDenied(state: State) {
-  return state.matches('bluetoothDenied');
-}
-
-export function selectIsLocationDenied(state: State) {
-  return state.matches('checkingLocationService.denied');
-}
-
-export function selectIsLocationDisabled(state: State) {
-  return state.matches('checkingLocationService.disabled');
-}
-
-export function selectIsDone(state: State) {
-  return state.matches('reviewing.navigatingToHome');
-}
-
-export function selectIsVerifyingIdentity(state: State) {
-  return state.matches('reviewing.verifyingIdentity');
-}
-
-export function selectIsInvalidIdentity(state: State) {
-  return state.matches('reviewing.invalidIdentity');
-}
-
-export function selectIsCancelling(state: State) {
-  return state.matches('reviewing.cancelling');
-}
-
-export function selectIsHandlingBleError(state: State) {
-  return state.matches('handlingBleError');
 }
