+++ screens/Scan/SendVcScreenController.ts
@@ -1,21 +1,23 @@
 import { useSelector } from '@xstate/react';
 import { useContext, useState } from 'react';
 import { ActorRefFrom } from 'xstate';
-import {
-  ScanEvents,
-  selectReason,
-  selectReceiverInfo,
-  selectIsSelectingVc,
-  selectVcName,
-  selectIsVerifyingIdentity,
-  selectIsInvalidIdentity,
-  selectSelectedVc,
-  selectIsCancelling,
-} from '../../machines/scan';
 import { selectVcLabel } from '../../machines/settings';
 import { selectShareableVcs } from '../../machines/vc';
 import { vcItemMachine } from '../../machines/vcItem';
 import { GlobalContext } from '../../shared/GlobalContext';
+import {
+  selectIsSelectingVc,
+  selectReason,
+  selectReceiverInfo,
+  selectSelectedVc,
+  selectVcName,
+} from '../../machines/openIdBle/scan/selectors';
+import {
+  selectIsCancelling,
+  selectIsInvalidIdentity,
+  selectIsVerifyingIdentity,
+} from '../../machines/openIdBle/commonSelectors';
+import { ScanEvents } from '../../machines/openIdBle/scan/machine';
 
 export function useSendVcScreen() {
   const { appService } = useContext(GlobalContext);
